Track single reference of done input selection offset
Previously, if editing an existing done then choosing a suggestion, a subsequent keypress would revert back to the original offset when editing started. This undesirable behavior can be avoided by treating selection offset as an initial value, and not resetting the suggestion offset. While the latter may be desirable, as long as we're not continuing to update reference, textarea won't update selection offset, and avoids a setState in componentDidUpdate